requirements.txt
openai-whisper>=20231117
openai>=1.3.0
python-dotenv>=1.0.0